IA Phase 2: Community hub — challenges live here now
Per MENU_IA_PLAN: Community becomes the "with people" hub and the standalone
Challenge Friends card is gone.

- Home cards are now Play · Community · Store.
- Community is a tabbed hub (default Challenges): Challenges · Leaderboard ·
  Activity, with a 👥 People button (Friends + Groups) in the header.
- The challenge inbox moved inline into the Challenges tab; the heavy builder
  stays as a focused "New Challenge" modal (stripped of its old inbox tab).
- Extracted LeaderboardPanel + ActivityPanel so the /leaderboard and /activity
  routes AND the Community tabs share one implementation (profile-panel pattern).
- Repointed callers: openChallenges() → Community ▸ Challenges; post-game
  "Challenge Friends" + the /?challenge deep-link → the New-Challenge builder.

Verified: QA 19/19, 0 console errors; live hub shows all three tabs rendering,
People (Friends/Groups), and the New-Challenge modal. Routes still work.
